requirements.txt
streamlit==1.29.0
streamlit-image-select==0.6.0
pandas==1.5.3
tensorflow==2.13.1
Pillow==9.0.1
matplotlib==3.4.2
numpy==1.23.1.
opencv-python==4.5.5.64
tensorflow==2.13.1
Pillow==9.0.1